requirements.txt
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
flashrank>=0.2.0
pdfplumber>=0.10.0
pymupdf>=1.23.0
pymupdf4llm>=0.0.10
